Renamed new images and removed old ones
