+++ client/src/components/Register.js
@@ -3,7 +3,7 @@ import axios from 'axios';
 import { registerThunk } from '../thunks';
 import { connect } from 'react-redux';
 import { Link } from 'react-router-dom';
-import logo from '../images/cst2.png';
+import logo from '../images/cunySecondBig.png';
 import '../styles/register.css';
 
 class Register extends Component {
